+++ config/services/commands.yml
@@ -11,3 +11,7 @@ services:
   PhpList\Core\Domain\Identity\Command\:
     resource: '../../src/Domain/Identity/Command'
     tags: ['console.command']
+
+  PhpList\Core\Domain\Messaging\Command\ProcessBouncesCommand:
+    arguments:
+      $protocolProcessors: !tagged_iterator 'phplist.bounce_protocol_processor'
