+++ config/services/processor.yml
@@ -1,0 +1,11 @@
+services:
+  _defaults:
+    autowire: true
+    autoconfigure: true
+    public: false
+
+  PhpList\Core\Domain\Messaging\Service\Processor\PopBounceProcessor:
+    tags: ['phplist.bounce_protocol_processor']
+
+  PhpList\Core\Domain\Messaging\Service\Processor\MboxBounceProcessor:
+    tags: ['phplist.bounce_protocol_processor']
