+++ src/Domain/Messaging/Command/ProcessBouncesCommand.php
@@ -9,7 +9,8 @@ use Exception;
 use PhpList\Core\Domain\Messaging\Model\Bounce;
 use PhpList\Core\Domain\Messaging\Model\UserMessage;
 use PhpList\Core\Domain\Messaging\Model\UserMessageBounce;
-use PhpList\Core\Domain\Messaging\Repository\MessageRepository;
+use PhpList\Core\Domain\Messaging\Service\BounceProcessingService;
+use PhpList\Core\Domain\Messaging\Service\Processor\BounceProtocolProcessor;
 use PhpList\Core\Domain\Messaging\Service\LockService;
 use PhpList\Core\Domain\Messaging\Service\Manager\BounceManager;
 use PhpList\Core\Domain\Messaging\Service\Manager\BounceRuleManager;
@@ -48,14 +49,15 @@ class ProcessBouncesCommand extends Command
 
     public function __construct(
         private readonly BounceManager $bounceManager,
-        private readonly SubscriberRepository $users,
-        private readonly MessageRepository $messages,
         private readonly BounceRuleManager $ruleManager,
         private readonly LockService $lockService,
         private readonly LoggerInterface $logger,
         private readonly SubscriberManager $subscriberManager,
         private readonly SubscriberHistoryManager $subscriberHistoryManager,
         private readonly SubscriberRepository $subscriberRepository,
+        private readonly BounceProcessingService $processingService,
+        /** @var iterable<BounceProtocolProcessor> */
+        private readonly iterable $protocolProcessors,
     ) {
         parent::__construct();
     }
@@ -82,61 +84,23 @@ class ProcessBouncesCommand extends Command
         try {
             $io->title('Processing bounces');
             $protocol = (string)$input->getOption('protocol');
-            $testMode = (bool)$input->getOption('test');
-            $max = (int)$input->getOption('maximum');
-            $purgeProcessed = $input->getOption('purge') && !$testMode;
-            $purgeUnprocessed = $input->getOption('purge-unprocessed') && !$testMode;
 
             $downloadReport = '';
 
-            if ($protocol === 'pop') {
-                $host = (string)$input->getOption('host');
-                $user = (string)$input->getOption('user');
-                $password = (string)$input->getOption('password');
-                $port = (string)$input->getOption('port');
-                $mailboxes = (string)$input->getOption('mailbox');
-
-                if (!$host || !$user || !$password) {
-                    $io->error('POP configuration incomplete: host, user, and password are required.');
-
-                    return Command::FAILURE;
+            $processor = null;
+            foreach ($this->protocolProcessors as $p) {
+                if ($p->getProtocol() === $protocol) {
+                    $processor = $p;
+                    break;
                 }
+            }
 
-                foreach (explode(',', $mailboxes) as $mailboxName) {
-                    $mailboxName = trim($mailboxName);
-                    if ($mailboxName === '') { $mailboxName = 'INBOX'; }
-                    $mailbox = sprintf('{%s:%s}%s', $host, $port, $mailboxName);
-                    $io->section("Connecting to $mailbox");
-
-                    $link = @imap_open($mailbox, $user, $password);
-                    if (!$link) {
-                        $io->error('Cannot create connection to '.$mailbox.': '.imap_last_error());
-
-                        return Command::FAILURE;
-                    }
-
-                    $downloadReport .= $this->processMessages($io, $link, $max, $purgeProcessed, $purgeUnprocessed, $testMode);
-                }
-            } elseif ($protocol === 'mbox') {
-                $file = (string)$input->getOption('mailbox');
-                if (!$file) {
-                    $io->error('mbox file path must be provided with --mailbox.');
-
-                    return Command::FAILURE;
-                }
-                $io->section("Opening mbox $file");
-                $link = @imap_open($file, '', '', $testMode ? 0 : CL_EXPUNGE);
-                if (!$link) {
-                    $io->error('Cannot open mailbox file: '.imap_last_error());
-
-                    return Command::FAILURE;
-                }
-                $downloadReport .= $this->processMessages($io, $link, $max, $purgeProcessed, $purgeUnprocessed, $testMode);
-            } else {
+            if ($processor === null) {
                 $io->error('Unsupported protocol: '.$protocol);
-
                 return Command::FAILURE;
             }
+
+            $downloadReport .= $processor->process($input, $io);
 
             // Reprocess unidentified bounces (status = "unidentified bounce")
             $this->reprocessUnidentified($io);
@@ -169,144 +133,6 @@ class ProcessBouncesCommand extends Command
         }
     }
 
-    private function processMessages(SymfonyStyle $io, $link, int $max, bool $purgeProcessed, bool $purgeUnprocessed, bool $testMode): string
-    {
-        $num = imap_num_msg($link);
-        $io->writeln(sprintf('%d bounces to fetch from the mailbox', $num));
-        if ($num === 0) {
-            imap_close($link);
-
-            return '';
-        }
-        $io->writeln('Please do not interrupt this process');
-        if ($num > $max) {
-            $io->writeln(sprintf('Processing first %d bounces', $max));
-            $num = $max;
-        }
-        $io->writeln($testMode ? 'Running in test mode, not deleting messages from mailbox' : 'Processed messages will be deleted from the mailbox');
-
-        for ($x = 1; $x <= $num; $x++) {
-            $header = imap_fetchheader($link, $x);
-            $processed = $this->processImapBounce($link, $x, $header, $io);
-            if ($processed) {
-                if (!$testMode && $purgeProcessed) {
-                    imap_delete($link, (string)$x);
-                }
-            } else {
-                if (!$testMode && $purgeUnprocessed) {
-                    imap_delete($link, (string)$x);
-                }
-            }
-        }
-
-        $io->writeln('Closing mailbox, and purging messages');
-        if (!$testMode) {
-            imap_close($link, CL_EXPUNGE);
-        } else {
-            imap_close($link);
-        }
-
-        return '';
-    }
-
-    private function processImapBounce($link, int $num, string $header, SymfonyStyle $io): bool
-    {
-        $headerInfo = imap_headerinfo($link, $num);
-        $date = $headerInfo->date ?? null;
-        $bounceDate = $date ? new DateTimeImmutable($date) : new DateTimeImmutable();
-        $body = imap_body($link, $num);
-        $body = $this->decodeBody($header, $body);
-
-        // Quick hack: ignore MsExchange delayed notices (as in original)
-        if (preg_match('/Action: delayed\s+Status: 4\.4\.7/im', $body)) {
-            return true;
-        }
-
-        $msgId = $this->findMessageId($body);
-        $userId = $this->findUserId($body);
-
-        $bounce = $this->bounceManager->create($bounceDate, $header, $body);
-
-        return $this->processBounceData($bounce, $msgId, $userId, $bounceDate);
-    }
-
-    private function processBounceData(
-        Bounce $bounce,
-        string|int|null $msgId,
-        ?int $userId,
-        DateTimeImmutable $bounceDate,
-    ): bool {
-        $msgId = $msgId ?: null;
-        if ($userId) {
-            $user = $this->subscriberManager->getSubscriberById($userId);
-        }
-
-        if ($msgId === 'systemmessage' && $userId) {
-            $this->bounceManager->update(
-                bounce: $bounce,
-                status: 'bounced system message',
-                comment: sprintf('%d marked unconfirmed', $userId))
-            ;
-            $this->bounceManager->linkUserMessageBounce($bounce,$bounceDate, $userId);
-            $this->subscriberManager->markUnconfirmed($userId);
-            $this->logger->info('system message bounced, user marked unconfirmed', ['userId' => $userId]);
-            $this->subscriberHistoryManager->addHistory(
-                subscriber: $user,
-                message: 'Bounced system message',
-                details: sprintf('User marked unconfirmed. Bounce #%d', $bounce->getId())
-            );
-
-            return true;
-        }
-
-        if ($msgId && $userId) {
-            if (!$this->bounceManager->existsUserMessageBounce($userId, (int)$msgId)) {
-                $this->bounceManager->linkUserMessageBounce($bounce, $bounceDate,$userId, (int)$msgId);
-                $this->bounceManager->update(
-                    bounce: $bounce,
-                    status: sprintf('bounced list message %d', $msgId),
-                    comment: sprintf('%d bouncecount increased', $userId)
-                );
-                $this->messages->incrementBounceCount((int)$msgId);
-                $this->users->incrementBounceCount($userId);
-            } else {
-                $this->bounceManager->linkUserMessageBounce($bounce, $bounceDate, $userId, (int)$msgId);
-                $this->bounceManager->update(
-                    bounce: $bounce,
-                    status: sprintf('duplicate bounce for %d', $userId),
-                    comment: sprintf('duplicate bounce for subscriber %d on message %d', $userId, $msgId)
-                );
-            }
-            return true;
-        }
-
-        if ($userId) {
-            $this->bounceManager->update(
-                bounce: $bounce,
-                status: 'bounced unidentified message',
-                comment: sprintf('%d bouncecount increased', $userId)
-            );
-            $this->users->incrementBounceCount($userId);
-            return true;
-        }
-
-        if ($msgId === 'systemmessage') {
-            $this->bounceManager->update($bounce, 'bounced system message', 'unknown user');
-            $this->logger->info('system message bounced, but unknown user');
-            return true;
-        }
-
-        if ($msgId) {
-            $this->bounceManager->update($bounce, sprintf('bounced list message %d', $msgId), 'unknown user');
-            $this->messages->incrementBounceCount((int)$msgId);
-            return true;
-        }
-
-        $this->bounceManager->update($bounce, 'unidentified bounce', 'not processed');
-
-        return false;
-    }
-
     private function reprocessUnidentified(SymfonyStyle $io): void
     {
         $io->section('Reprocessing unidentified bounces');
@@ -319,12 +145,12 @@ class ProcessBouncesCommand extends Command
             if ($count % 25 === 0) {
                 $io->writeln(sprintf('%d out of %d processed', $count, $total));
             }
-            $decodedBody = $this->decodeBody($bounce->getHeader(), $bounce->getData());
-            $userId = $this->findUserId($decodedBody);
-            $messageId = $this->findMessageId($decodedBody);
+            $decodedBody = $this->processingService->decodeBody($bounce->getHeader(), $bounce->getData());
+            $userId = $this->processingService->findUserId($decodedBody);
+            $messageId = $this->processingService->findMessageId($decodedBody);
             if ($userId || $messageId) {
                 $reparsed++;
-                if ($this->processBounceData($bounce->getId(), $messageId, $userId, new DateTimeImmutable())) {
+                if ($this->processingService->processBounceData($bounce, $messageId, $userId, new DateTimeImmutable())) {
                     $reidentified++;
                 }
             }
@@ -488,7 +314,7 @@ class ProcessBouncesCommand extends Command
                         break;
                     }
                 }
-                if ($removed || $msgokay) {
+                if ($removed) {
                     break;
                 }
             }
@@ -499,47 +325,4 @@ class ProcessBouncesCommand extends Command
         $io->writeln(sprintf('total of %d subscribers processed', $total));
     }
 
-    private function decodeBody(string $header, string $body): string
-    {
-        $transferEncoding = '';
-        if (preg_match('/Content-Transfer-Encoding: ([\w-]+)/i', $header, $regs)) {
-            $transferEncoding = strtolower($regs[1]);
-        }
-        return match ($transferEncoding) {
-            'quoted-printable' => quoted_printable_decode($body),
-            'base64' => base64_decode($body) ?: '',
-            default => $body,
-        };
-    }
-
-    private function findMessageId(string $text): string|int|null
-    {
-        if (preg_match('/(?:X-MessageId|X-Message): (.*)\r\n/iU', $text, $match)) {
-            return trim($match[1]);
-        }
-        return null;
-    }
-
-    private function findUserId(string $text): ?int
-    {
-        // Try X-ListMember / X-User first
-        if (preg_match('/(?:X-ListMember|X-User): (.*)\r\n/iU', $text, $match)) {
-            $user = trim($match[1]);
-            if (str_contains($user, '@')) {
-                return $this->users->idByEmail($user);
-            } elseif (preg_match('/^\d+$/', $user)) {
-                return (int)$user;
-            } elseif ($user !== '') {
-                return $this->users->idByUniqId($user);
-            }
-        }
-        // Fallback: parse any email in the body and see if it is a subscriber
-        if (preg_match_all('/[._a-zA-Z0-9-]+@[.a-zA-Z0-9-]+/', $text, $regs)) {
-            foreach ($regs[0] as $email) {
-                $id = $this->users->idByEmail($email);
-                if ($id) { return $id; }
-            }
-        }
-        return null;
-    }
 }
